@@ -5,50 +5,41 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
+    "build": "ng build --configuration production",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "dependencies": {
-    "@angular/animations": "~11.0.0",
-    "@angular/common": "~11.0.0",
-    "@angular/compiler": "~11.0.0",
-    "@angular/core": "~11.0.0",
-    "@angular/forms": "~11.0.0",
-    "@angular/platform-browser": "~11.0.0",
-    "@angular/platform-browser-dynamic": "~11.0.0",
-    "@angular/router": "~11.0.0",
-    "@ngrx/effects": "^10.0.1",
-    "@ngrx/entity": "^10.0.1",
-    "@ngrx/store": "^10.0.1",
-    "@ngrx/store-devtools": "^10.0.1",
-    "@ngrx/data": "^10.0.1",
-    "angular-in-memory-web-api": "^0.11.0",
+    "@angular/animations": "^14.1.2",
+    "@angular/common": "^14.1.2",
+    "@angular/compiler": "^14.1.2",
+    "@angular/core": "^14.1.2",
+    "@angular/forms": "^14.1.2",
+    "@angular/platform-browser": "^14.1.2",
+    "@angular/platform-browser-dynamic": "^14.1.2",
+    "@angular/router": "^14.1.2",
+    "@ngrx/effects": "^14.1.0",
+    "@ngrx/entity": "^14.1.0",
+    "@ngrx/store": "^14.1.0",
+    "@ngrx/store-devtools": "^14.1.0",
+    "@ngrx/data": "^14.1.0",
+    "angular-in-memory-web-api": "^0.14.0",
     "core-js": "^2.5.4",
     "rxjs": "^6.6.3",
     "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.1",
-    "@angular/cli": "~11.0.1",
-    "@angular/compiler-cli": "11.0.0",
-    "@angular/language-service": "11.0.0",
+    "@angular-devkit/build-angular": "^14.1.2",
+    "@angular/cli": "^14.1.2",
+    "@angular/compiler-cli": "14.1.2",
     "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~4.1.0",
-    "tslint": "~6.1.0",
-    "typescript": "4.0.5"
+    "typescript": "4.6.4"
   }
 }
